requirements.txt
# Bot dependencies
discord.py>=2.3,<3.0
aiohttp>=3.9
python-dotenv>=1.0
psutil>=5.9
PyNaCl>=1.5.0
